+++ vite.config.ts
@@ -39,9 +39,18 @@ export default defineConfig({
             }
         }
     },
-    // Force optimization of UiKit in dev mode to avoid to avoid the error:
+    // Force optimization of UiKit and limax (not module packages) in dev mode 
+    // to avoid the error:
     // "importing binding name 'default' cannot be resolved by star export entries"
+    // Also, treat all sprinkles as source code (not prebuilt) and tell Vite 
+    // not to prebundle them.
     optimizeDeps: {
-        include: ['uikit', 'uikit/dist/js/uikit-icons', 'limax']
+        include: ['uikit', 'uikit/dist/js/uikit-icons', 'limax'],
+        exclude: [
+            '@userfrosting/sprinkle-core',
+            '@userfrosting/sprinkle-account',
+            '@userfrosting/sprinkle-admin',
+            '@userfrosting/theme-pink-cupcake'
+        ]
     }
 })